requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask-sqlalchemy==2.5.1
flask-wtf==0.14.3
python-dateutil==2.6.0
jinja2==2.0.0
Werkzeug==3.0
psycop2-binary==2.9.3
WTForms==3.0.0

